changes made to Product

+++ src/main/java/com/inventory/management/dto/ProductRequest.java
@@ -6,8 +6,6 @@ import jakarta.validation.constraints.Positive;
 import jakarta.validation.constraints.PositiveOrZero;
 import lombok.Data;
 
-import java.math.BigDecimal;
-
 @Data
 public class ProductRequest {
     
@@ -21,7 +19,7 @@ public class ProductRequest {
     
     @NotNull(message = "Price is required")
     @Positive(message = "Price must be positive")
-    private BigDecimal price;
+    private float price;
     
     @NotNull(message = "Quantity is required")
     @PositiveOrZero(message = "Quantity must be zero or positive")
